Working implementation of insert and put
Fixed lots of bugs, and got the insert and put methods working.
put() needed its own XS implementation because it returns the
previous value for the specific key, and that was best done in
XS.  I also have it remove all existing values for a key if the
tree allows duplicate keys.