Upgrade enum constants to dualvars, so users can see what the value is
Enums are nice for C code but suck in Perl code when it just
shows an integer, leaving you guessing about the constant.
Now the KEY_TYPE_* constants are dualvars that stringify usefully
but also can be passed to an XS function that wants integers.
Also, allow plain strings to be given to the constructor
and coerce them to the constant value.

Also more documentation improvements.