Refactor to show user the compare fn and support optional wantarray
I'm not willing to have get() change its behavior in list context by default,
but added an option to enable the Tree::RB behavior for anyone who needs it.

I also restructured the initialization so that the compare function is
processed as an enum, allowing more custom compare functions to be added.