Convert iterators to XS, and implement half of TIEHASH interface
In the process of implementing TIEHASH, I decided to see if I could
enable iterating the tree without inflating each and every node.
The new iterator structs point to an _item and move to the next
(or previous) item any time the item is removed from the tree.
The iterator now has 'key' and 'value' attributes so that the node
is not required, and calling 'next' in void context moves to the
next node without creating a node object.