Fix unicode comparison bugs
I discovered that I misunderstood the putpose of bytes_cmp_utf8. It is not
a utf8 strcmp, it is a compare-nonutf8-with-utf8 function for the purpose
of not having to convert the bytes of one side of the cmp operator.

With that, I studied the Perl_sv_cmp code and realized that they are just
calling memcmp on the unicode or bytes as long as both strings are the same
encoding. (except for ebcdic which involves insanity that I'm not going to
mess with) so I did likewise.  On Perl 5.14 and later, I use SvPV for
whatever encoding is available, and bytes_cmp_utf8, and before 5.14, I
decode both SVs as bytes or as utf8 so that memcmp works correctly.