Implement put_multi, insert_multi, and optimization for insert locality
In the process of writing put_multi and insert_multi, I decided that
there ought to be an optimization for inserting a pre-sorted list
in sequential order with N comparisons instead of N*log(N).
After several attempts, I came up with the idea that it could just be
triggered by insertions being adjacent to previous insertions some
number of times in a row, which would make it worthwhile to run one
or two additional comparisons to see if that trend is being continued.

I also refactored all the insertion code into a single method, and it
uses the absolute minimum number of comparisons now, instead of the
rbtree_insert(hint...) that had to verify the hint during the
insertion.