Implement new constructor in XS, for deserialization
This speeds up the regular constructor, and also allows more of
the state of the tree to be initialized; new attributes are 'keys',
'values', 'recent', and 'hashiter'.

This also implements most of a STORABLE_thaw but I misread the docs
and didn't realize all the rest of the arguments needed to be
references.  I'll need to wrap up the arguments, but maybe also
take the opportunity to pack some of them into bytes.

refs #3