Fix various crashes in node methods when removed from tree
When the tree is NULL, I need to be sure not to access it.
Unit tests also reminded me that count becomes zero, so I
documented that.  I had also forgotten that nodes don't
hold a strong ref to the tree, so I documented that too
and added a test case.